DVC-1887 - Add new public key for UniFHIR (#3)
# Description

https://doccla.atlassian.net/browse/DVC-1887

Authentication is not currently working using the `unifhir-non-prod-1`
key.
We suspect something is wrong with the key, maybe the public/private
keys don't match.

This PR adds another UniFHIR key into the non-prod keyset to test this
theory. If this keys works, there will be a followup PR to this to
remove the `unifhir-non-prod-1` key and create a new prod one too.

The key pair has been stored in 1password under the label
'unifhir-non-prod-2 identity keys'

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

- [x] Manually

Wrote a temporary test using the `buddy/sign` library to verify we can
`sign` and `unsign` a JWT using these keys. The test works for
`unifhir-non-prod-2` but not for `unifhir-non-prod-1`.