DVC-1979 - Add new production public key for UniFHIR (#4)
# Description

https://doccla.atlassian.net/browse/DVC-1979

When testing with MFT (Epic customer) we discovered the non-prod
key-pair wasn't working to verify the JWT on their side. After some
debugging we just generated a new key (unifhir-non-prod-2) and that
worked fine.

This is to also update the prod key so we can have more confidence the
keys will work when we get to production too.

There will be an accompanying UniFHIR PR to update the config to use
these keys and also the private key GCP secrets will get updated for the
new keys.

The keys are stored in the 1password entry 'unifhir-prod-2 identity
keys'

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

- [x] Manually

Verified the non-prod-2 key works with the MFT test environment from a
local stack. This is an identical change for prod so it should work too
🤞